big update

+++ src/main/java/by/golik/dealerstat/controller/UserController.java
@@ -1,22 +1,23 @@
 package by.golik.dealerstat.controller;
 
-import by.golik.dealerstat.entity.Role;
 import by.golik.dealerstat.entity.User;
+import by.golik.dealerstat.exception.NotEnoughRightException;
 import by.golik.dealerstat.exception.ResourceNotFoundException;
 import by.golik.dealerstat.service.GameObjectService;
 import by.golik.dealerstat.service.UserService;
+import by.golik.dealerstat.service.dto.RoleDTO;
+import by.golik.dealerstat.service.dto.UserDTO;
+import by.golik.dealerstat.service.util.UserDtoAssembler;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import javax.validation.ValidationException;
 import javax.validation.constraints.DecimalMax;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
 import java.security.Principal;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -36,49 +37,32 @@ public class UserController {
         this.gameObjectService = gameObjectService;
     }
 
-    //todo
-    @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<User> get(@PathVariable int id) throws ResourceNotFoundException {
-        Principal principal = (Principal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    @GetMapping("/{id}")
+    public UserDTO getUser(@PathVariable("id") int id, Principal principal) throws ResourceNotFoundException {
         User user = userService.getUser(id);
         User principalUser = userService.getUserByEmailAndEnabled(principal.getName());
+        UserDTO userDTO;
+
         userService.calculateRate(user);
-
+        userDTO = UserDtoAssembler.toDto(user);
         if (userService.isAdmin(principalUser) || principalUser.getId() == id) {
-            user.setEmail(user.getEmail());
+            userDTO.setEmail(user.getEmail());
         }
-        if (!this.userService.findByUserId(id).isPresent()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>(user, HttpStatus.OK);
+        return userDTO;
     }
 
-    /**
-     *
-     * @return
-     */
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<User>> getAllUsers() {
-        List<User> users = this.userService.getAllUsers();
-        return generateListResponse(users);
+    @GetMapping
+    public List<UserDTO> getAllUsers() {
+        return UserDtoAssembler.toDtoList(userService.getAllUsers());
     }
 
-    /**
-     * norm
-     * @return
-     */
-    @GetMapping(value = "/readers",produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<User>> getAllReaders() {
-        List<User> users = this.userService.getAllAnons();
-        return generateListResponse(users);
+    @GetMapping("/readers")
+    public List<UserDTO> getAllReaders() {
+        return UserDtoAssembler.toDtoList(userService.getAllAnons());
     }
 
-    /**
-     * norm
-     * @return
-     */
     @GetMapping("/traders")
-    public ResponseEntity<List<User>> getAllTraders(@RequestParam(required = false, defaultValue = "5")
+    public List<UserDTO> getAllTraders(@RequestParam(required = false, defaultValue = "5")
                                        @DecimalMax(value = "5.0",
                                                message = "min parameter should be less or equal than 5.0") double max,
                                        @RequestParam(required = false, defaultValue = "1") @DecimalMin(value = "1.0",
@@ -88,39 +72,61 @@ public class UserController {
                                        @Min(value = 0, message = "skip parameter should be more or equal than 0") int skip,
                                        @RequestParam(required = false, defaultValue = "0")
                                        @Min(value = 0, message = "limit parameter should be more or equal than 0") int limit) {
-        List<User> users = this.userService.getAllTraders();
-        return generateListResponse(users);
-    }
+        String[] split;
+        List<Integer> idList;
+        List<User> users;
 
-
-    @PutMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<User> updateGame(@RequestBody User user, @PathVariable int id) {
-        return new ResponseEntity<>(userService.update(user, id));
-    }
-
-    @PatchMapping(value = "/{id}/change-role", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<User> changeRole(@PathVariable("id") int id,
-                           @RequestBody @Valid Role role) throws ResourceNotFoundException {
-        User user = userService.getUser(id);
-        return new ResponseEntity<>(userService.changeRole(user, role.getName()));
-    }
-
-    /**
-     * if not enable - don't find this user
-     * @param id
-     * @return
-     */
-    @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<User> deleteUser(@PathVariable("id") int id) throws ResourceNotFoundException {
-        User user = userService.getUser(id);
-        userService.deleteUser(user);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    protected ResponseEntity<List<User>> generateListResponse(List<User> users) {
-        if (users.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (min > max) {
+            throw new ValidationException("min parameter should be less than max!");
         }
-        return new ResponseEntity<>(users, HttpStatus.OK);
+        try {
+            split = games.split(",");
+            idList = gameObjectService.getGameIdByName(split);
+            if (idList == null) {
+                return new ArrayList<UserDTO>();
+            }
+            users = userService.getAllTradersByGames(idList);
+        }
+        catch (NullPointerException e) {
+            users = userService.getAllTraders();
+            System.out.println(users);
+        }
+//        users = userService.filterTraders(users, max, min, skip, limit);
+        return UserDtoAssembler.toDtoList(users);
+    }
+
+    @PutMapping("/my")
+    public void updateUser(@RequestBody @Valid UserDTO userDTO,
+                           Principal principal) throws ResourceNotFoundException {
+        User user = userService.getUserByEmailAndEnabled(principal.getName());
+
+        userService.updateUser(user, userDTO);
+    }
+
+    @PatchMapping("/{id}/change-role")
+    public void changeRole(@PathVariable("id") int id,
+                           @RequestBody @Valid RoleDTO roleDTO) throws ResourceNotFoundException,
+            NotEnoughRightException {
+        User user = userService.getUser(id);
+
+        if (userService.isAdmin(user)) {
+            throw new NotEnoughRightException("You can't change role of admin!");
+        }
+        userService.changeRole(user, roleDTO.getRole());
+    }
+
+    @DeleteMapping("/{id}")
+    public void deleteUser(@PathVariable("id") int id, Principal principal) throws ResourceNotFoundException,
+            NotEnoughRightException {
+        User user = userService.getUser(id);
+        User principalUser = userService.getUserByEmailAndEnabled(principal.getName());
+
+        if (!userService.isAdmin(principalUser) && !principalUser.equals(user)) {
+            throw new NotEnoughRightException("You can't delete this user!");
+        }
+        if (userService.isAdmin(user)) {
+            throw new NotEnoughRightException("You can't delete admin!");
+        }
+        userService.deleteUser(user);
     }
 }
